refactor: Reorganize build process
