+++ packages/aws-lambda-otel-extension-dist/.gitignore
@@ -1,2 +1,1 @@
 /*.zip
-/package.json
