+++ packages/aws-lambda-otel-extension-dist/.npmignore
@@ -1,0 +1,1 @@
+/scripts
