+++ packages/aws-lambda-otel-extension-dist/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@serverless/aws-lambda-otel-extension-dist",
+  "author": "Serverless, Inc.",
+  "devDependencies": {
+    "@serverless/aws-lambda-otel-extension": "^0.0.0",
+    "essentials": "^1.2.0",
+    "fs2": "^0.3.9"
+  },
+  "scripts": {
+    "build": "./scripts/build.js",
+    "prepare-package-json": "./scripts/prepare-package-json.js"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT"
+}
