+++ packages/aws-lambda-otel-extension-dist/scripts/build.js
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+
+'use strict';
+
+require('essentials');
+
+require('./lib/build')();
