+++ packages/aws-lambda-otel-extension-dist/scripts/prepare-package-json.js
@@ -4,4 +4,4 @@
 
 require('essentials');
 
-require('./build')();
+require('./lib/prepare-package-json')();
