+++ packages/aws-lambda-otel-extension/.gitignore
@@ -1,2 +1,1 @@
 /dist
-/test
