+++ packages/aws-lambda-otel-extension/README.md
@@ -1,5 +1,5 @@
 # @serverless/aws-lambda-otel-extension
 
-AWS Lambda extension that gathers otel telemetry and sends to designated url.
+## AWS Lambda extension that gathers otel telemetry and sends to designated url.
 
 _Note: Package it's still in experimental stage and subject to major changes._
