+++ packages/aws-lambda-otel-extension/package.json
@@ -2,7 +2,8 @@
   "name": "@serverless/aws-lambda-otel-extension",
   "repository": "serverless/runtime",
   "version": "0.0.0",
-  "devDependencies": {
+  "author": "Serverless, Inc.",
+  "dependencies": {
     "adm-zip": "^0.5.9",
     "child-process-ext": "^2.1.1",
     "esbuild": "^0.14.21",
@@ -11,7 +12,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "build": "./scripts/build/index.js",
+    "build": "./scripts/build.js",
     "prebuild:prod": "esbuild opt/otel-extension/index.js --bundle --platform=node --target=node12 --minify --outfile=opt/build/index.js && esbuild opt/otel-extension/wrapper.js --bundle --minify --platform=node --target=node12 --outfile=opt/build/wrapper.js",
     "build:prod": "./scripts/build.prod.js",
     "test": "mocha \"test/unit/**/*.test.js\"",
